Abstraction in RenderModel + normals per triangle

+++ core/Vertex.h
@@ -8,7 +8,7 @@
 #include <glm/glm.hpp>
 #include <ostream>
 
-typedef glm::dvec3 Vertex;
+typedef glm::vec3 Vertex;
 std::ostream & operator<<(std::ostream & os, const Vertex& v);
 
 #endif //OPTIX_SAMPLES_VERTEX_H
